refactor: API 호출 함수 try-catch문 제거 및 훅 내 선언으로 수정
