+++ src/queries/chat/useChatListQuery.ts
@@ -1,5 +1,14 @@
 import { useQuery } from "@tanstack/react-query";
-import { fetchChatList } from "@/apis/sidebar/fetchChatList";
+import axiosInstance from "@/apis/utils/axiosInstance";
+
+type FetchChatListPayload = {
+  searchText: string;
+};
+
+export const fetchChatList = async ({ searchText }: FetchChatListPayload) => {
+  const response = await axiosInstance.post("/chats", { searchText });
+  return response.data;
+};
 
 export const useChatListQuery = (searchText: string) => {
   return useQuery({
