+++ src/queries/chat/useCreateChatMutation.ts
@@ -1,5 +1,10 @@
 import { useMutation } from "@tanstack/react-query";
-import { postChat } from "@/apis/chat/postChat";
+import axiosInstance from "@/apis/utils/axiosInstance";
+
+export const postChat = async () => {
+  const response = await axiosInstance.post("/chats/new");
+  return response.data;
+};
 
 export const useCreateChatMutation = () =>
   useMutation({
