+++ src/queries/filter/useSetFilterMutation.ts
@@ -1,9 +1,14 @@
 import { useMutation } from "@tanstack/react-query";
-import { postChatFilterSettings } from "@/apis/filter/postChatFilterSettings";
+import axiosInstance from "@/apis/utils/axiosInstance";
 import { FilterPayload } from "@/types/filter";
 
-export const usePostChatFilterSettingsMutation = () => {
+export const postChatFilter = async (payload: FilterPayload) => {
+  const response = await axiosInstance.post("/chats/filters", payload);
+  return response.data;
+};
+
+export const useSetFilterMutation = () => {
   return useMutation({
-    mutationFn: (payload: FilterPayload) => postChatFilterSettings(payload),
+    mutationFn: (payload: FilterPayload) => postChatFilter(payload),
   });
 };
